AI research agent/requirements.txt
langchain>=0.2.0
langchain-groq>=0.1.0
langchain-core>=0.2.0
langsmith>=0.1.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.2
pydantic>=2.5.0
